requirements.txt
# Core LLM & agent frameworks
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
crewai>=0.80.0
crewai-tools>=0.14.0

# Observability
langsmith>=0.1.0

# Orchestration
prefect>=3.0.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0

# Frontend
streamlit>=1.38.0

# Data & storage
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.5.0
rich>=13.8.0

# Dev / testing
pytest>=8.3.0
pytest-asyncio>=0.23.0
